+++ code/crates/core-consensus/src/handle/proposal.rs
@@ -1,6 +1,5 @@
 use crate::handle::driver::apply_driver_input;
 use crate::handle::signature::verify_signature;
-use crate::handle::validator_set::get_validator_set;
 use crate::input::Input;
 use crate::prelude::*;
 use crate::types::{ConsensusMsg, ProposedValue, SignedConsensusMsg, WalEntry};
@@ -36,8 +35,7 @@ pub async fn on_proposal<Ctx>(
 where
     Ctx: Context,
 {
-    let consensus_height = state.driver.height();
-
+    let consensus_height = state.height();
     let proposal_height = signed_proposal.height();
     let proposal_round = signed_proposal.round();
     let proposer_address = signed_proposal.validator_address();
@@ -148,21 +146,14 @@ pub async fn verify_signed_proposal<Ctx>(
 where
     Ctx: Context,
 {
-    let consensus_height = state.driver.height();
+    let consensus_height = state.height();
     let proposal_height = signed_proposal.height();
     let proposal_round = signed_proposal.round();
     let proposer_address = signed_proposal.validator_address();
 
-    let Some(validator_set) = get_validator_set(co, state, proposal_height).await? else {
-        debug!(
-            consensus.height = %consensus_height,
-            proposal.height = %proposal_height,
-            proposer = %proposer_address,
-            "Received proposal for height without known validator set, dropping"
-        );
+    assert_eq!(proposal_height, consensus_height);
 
-        return Ok(false);
-    };
+    let validator_set = state.validator_set();
 
     let Some(proposer) = validator_set.get_by_address(proposer_address) else {
         warn!(